[Minor]  Moving Argot code under Obscurers
Minor re-organization - Moving Argot code under Obscurers
